[tests][s]: moved to mocha to test on travis - refs #8
